requirements.txt
pillow
pygame
pymupdf
requests
schedule
